tests/test_role_endpoints.py
"""
Tests for role endpoints.
"""

import pytest
from uuid import uuid4
from fastapi import status


@pytest.mark.unit
def test_list_roles_without_auth(test_client):
    """Test listing roles without authentication."""
    response = test_client.get("/api/v1/authz/roles")
    assert response.status_code == status.HTTP_401_UNAUTHORIZED


@pytest.mark.unit
def test_list_roles_with_auth(test_client, auth_headers):
    """Test listing roles with authentication."""
    response = test_client.get(
        "/api/v1/authz/roles",
        headers=auth_headers
    )
    # Since entity service is not running, expect error or empty list
    assert response.status_code in [status.HTTP_200_OK, status.HTTP_500_INTERNAL_SERVER_ERROR]


@pytest.mark.unit
def test_create_role_invalid_data(test_client, auth_headers):
    """Test creating role with invalid data."""
    response = test_client.post(
        "/api/v1/authz/roles",
        json={
            "name": "ab",  # Too short (min 3 chars)
            "tenant_id": str(uuid4())
        },
        headers=auth_headers
    )
    assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY


@pytest.mark.unit
def test_get_role_not_found(test_client, auth_headers):
    """Test getting non-existent role."""
    role_id = uuid4()
    response = test_client.get(
        f"/api/v1/authz/roles/{role_id}",
        headers=auth_headers
    )
    # Since entity service is not running, expect 404 or 500
    assert response.status_code in [status.HTTP_404_NOT_FOUND, status.HTTP_500_INTERNAL_SERVER_ERROR]


@pytest.mark.unit
def test_list_roles_with_pagination(test_client, auth_headers):
    """Test listing roles with pagination parameters."""
    response = test_client.get(
        "/api/v1/authz/roles?page=1&page_size=10",
        headers=auth_headers
    )
    assert response.status_code in [status.HTTP_200_OK, status.HTTP_500_INTERNAL_SERVER_ERROR]


@pytest.mark.unit
def test_list_roles_with_tenant_filter(test_client, auth_headers, sample_tenant_id):
    """Test listing roles filtered by tenant."""
    response = test_client.get(
        f"/api/v1/authz/roles?tenant_id={sample_tenant_id}",
        headers=auth_headers
    )
    assert response.status_code in [status.HTTP_200_OK, status.HTTP_500_INTERNAL_SERVER_ERROR]


@pytest.mark.unit
def test_update_role_partial(test_client, auth_headers, sample_role_id):
    """Test updating role with partial data."""
    response = test_client.patch(
        f"/api/v1/authz/roles/{sample_role_id}",
        json={"description": "Updated description"},
        headers=auth_headers
    )
    # Since entity service is not running, expect error
    assert response.status_code in [status.HTTP_200_OK, status.HTTP_404_NOT_FOUND, status.HTTP_500_INTERNAL_SERVER_ERROR]


@pytest.mark.unit
def test_delete_role(test_client, auth_headers, sample_role_id):
    """Test deleting role."""
    response = test_client.delete(
        f"/api/v1/authz/roles/{sample_role_id}",
        headers=auth_headers
    )
    # Since entity service is not running, expect error
    assert response.status_code in [status.HTTP_204_NO_CONTENT, status.HTTP_404_NOT_FOUND, status.HTTP_500_INTERNAL_SERVER_ERROR]
